+++ tests/test-urls.jasmine.ts
@@ -1,37 +1,21 @@
-import fs from "fs";
-import os from "os";
-import path from "path";
-import rimraf from "rimraf";
-import slugify from "slugify";
-import dotenv from "dotenv";
 import { Storage } from "../src/Storage";
-import to from "await-to-js";
 import "jasmine";
-import {
-  IStorage,
-  StorageConfig,
-  StorageType,
-  ConfigLocal,
-  ConfigGoogleCloud,
-  ConfigAmazonS3,
-} from "../src/types";
-dotenv.config();
+import { StorageType } from "../src/types";
 
 const urlsGoogle = [
   "gcs://keyFile.json:appName/the-buck",
   "gcs://keyFile.json:appName/",
   "gcs://keyFile.json:appName",
+  "gcs://tests/keyFile.json/the-buck",
   "gcs://tests/keyFile.json/",
   "gcs://keyFile.json",
 ];
 
 const urlsAmazon = [
-  "s3://key:secret@eu-west-2/the-buck",
-  "s3://key:secret@eu-west-2/",
-  "s3://key:secret@eu-west-2",
-  "s3://key:secret@eu-west-2/the-buck?sslEnabled=true",
-  "s3://key:secret@/the-buck",
-  "s3://key:secret@/the-buck",
+  "s3://key:secret",
+  "s3://key:secret?region=eu-west-2&bucketName=the-buck&sslEnabled=true",
+  "s3://key:secret?buckeName=the-buck",
+  "s3://key:secret?region=eu-west-2&bucketName=the-buck&sslEnabled=true&useDualstack=23&nonExistentKey=true",
 ];
 const urlsLocal = ["local://tests/tmp/the-buck", "local://tests/tmp", ""];
 
@@ -43,7 +27,7 @@ describe(`testing Google urls`, () => {
   });
   beforeAll(() => {});
 
-  it("url 0 config", () => {
+  it("[0]", () => {
     this.storage = new Storage(urlsGoogle[i]);
     expect(this.storage.introspect("type")).toBe(StorageType.GCS);
     expect(this.storage.introspect("bucketName")).toBe("the-buck");
@@ -51,7 +35,7 @@ describe(`testing Google urls`, () => {
     expect(this.storage.introspect("keyFilename")).toBe("keyFile.json");
   });
 
-  it("url 1 config", () => {
+  it("[1]", () => {
     this.storage = new Storage(urlsGoogle[i]);
     expect(this.storage.introspect("type")).toBe(StorageType.GCS);
     expect(this.storage.introspect("bucketName")).toBe("");
@@ -59,7 +43,7 @@ describe(`testing Google urls`, () => {
     expect(this.storage.introspect("keyFilename")).toBe("keyFile.json");
   });
 
-  it("url 2 config", () => {
+  it("[2]", () => {
     this.storage = new Storage(urlsGoogle[i]);
     expect(this.storage.introspect("type")).toBe(StorageType.GCS);
     expect(this.storage.introspect("bucketName")).toBe("");
@@ -67,11 +51,103 @@ describe(`testing Google urls`, () => {
     expect(this.storage.introspect("keyFilename")).toBe("keyFile.json");
   });
 
-  it("url 3 config", () => {
+  it("[3]", () => {
     this.storage = new Storage(urlsGoogle[i]);
     expect(this.storage.introspect("type")).toBe(StorageType.GCS);
-    expect(this.storage.introspect("bucketName")).toBe("");
-    expect(this.storage.introspect("projectId")).toBe("appIdFromFile");
+    expect(this.storage.introspect("bucketName")).toBe("the-buck");
+    expect(this.storage.introspect("projectId")).toBe("appIdFromTestFile");
     expect(this.storage.introspect("keyFilename")).toBe("tests/keyFile.json");
+  });
+
+  it("[4]", () => {
+    this.storage = new Storage(urlsGoogle[i]);
+    expect(this.storage.introspect("type")).toBe(StorageType.GCS);
+    expect(this.storage.introspect("bucketName")).toBe("");
+    expect(this.storage.introspect("projectId")).toBe("appIdFromTestFile");
+    expect(this.storage.introspect("keyFilename")).toBe("tests/keyFile.json");
+  });
+
+  it("[5]", () => {
+    this.storage = new Storage(urlsGoogle[i]);
+    expect(this.storage.introspect("type")).toBe(StorageType.GCS);
+    expect(this.storage.introspect("bucketName")).toBe("");
+    expect(this.storage.introspect("projectId")).toBe("appIdFromTestFile");
+    expect(this.storage.introspect("keyFilename")).toBe("keyFile.json");
+  });
+});
+
+describe(`testing Amazon urls`, () => {
+  let i = 0;
+  afterEach(() => {
+    i++;
+  });
+  beforeAll(() => {});
+
+  it("[0]", () => {
+    this.storage = new Storage(urlsAmazon[i]);
+    expect(this.storage.introspect("type")).toBe(StorageType.S3);
+    expect(this.storage.introspect("accessKeyId")).toBe("key present but hidden");
+    expect(this.storage.introspect("secretAccessKey")).toBe("secret present but hidden");
+    expect(this.storage.introspect("region")).toBe(undefined);
+    expect(this.storage.introspect("bucketName")).toBe(undefined);
+  });
+
+  it("[1]", () => {
+    this.storage = new Storage(urlsAmazon[i]);
+    expect(this.storage.introspect("type")).toBe(StorageType.S3);
+    expect(this.storage.introspect("accessKeyId")).toBe("key present but hidden");
+    expect(this.storage.introspect("secretAccessKey")).toBe("secret present but hidden");
+    expect(this.storage.introspect("region")).toBe("eu-west-2");
+    expect(this.storage.introspect("bucketName")).toBe("the-buck");
+  });
+
+  it("[2] typo", () => {
+    this.storage = new Storage(urlsAmazon[i]);
+    expect(this.storage.introspect("type")).toBe(StorageType.S3);
+    expect(this.storage.introspect("accessKeyId")).toBe("key present but hidden");
+    expect(this.storage.introspect("secretAccessKey")).toBe("secret present but hidden");
+    expect(this.storage.introspect("region")).toBe(undefined);
+    expect(this.storage.introspect("bucketName")).toBe(undefined);
+  });
+
+  it("[3] non-existent keys and key with wrong value types", () => {
+    this.storage = new Storage(urlsAmazon[i]);
+    expect(this.storage.introspect("type")).toBe(StorageType.S3);
+    expect(this.storage.introspect("accessKeyId")).toBe("key present but hidden");
+    expect(this.storage.introspect("secretAccessKey")).toBe("secret present but hidden");
+    expect(this.storage.introspect("region")).toBe("eu-west-2");
+    expect(this.storage.introspect("bucketName")).toBe("the-buck");
+    expect(this.storage.introspect("sslEnabled")).toBe(true);
+    expect(this.storage.introspect("useDualStack")).toBe(undefined);
+    expect(this.storage.introspect("nonExistentKey")).toBe(undefined);
+  });
+});
+
+describe(`testing local urls`, () => {
+  let i = 0;
+  afterEach(() => {
+    i++;
+  });
+  beforeAll(() => {});
+
+  it("[0]", () => {
+    this.storage = new Storage(urlsLocal[i]);
+    expect(this.storage.introspect("type")).toBe(StorageType.LOCAL);
+    expect(this.storage.introspect("directory")).toBe("tests/tmp/");
+    expect(this.storage.introspect("bucketName")).toBe("the-buck");
+  });
+
+  it("[1]", () => {
+    this.storage = new Storage(urlsLocal[i]);
+    expect(this.storage.introspect("type")).toBe(StorageType.LOCAL);
+    expect(this.storage.introspect("directory")).toBe("tests/");
+    expect(this.storage.introspect("bucketName")).toBe("tmp");
+  });
+
+  it("[2] no config", () => {
+    this.storage = new Storage(urlsLocal[i]);
+    expect(this.storage.introspect("type")).toBe(StorageType.LOCAL);
+    expect(this.storage.introspect("directory")).toBe("/tmp/");
+    expect(this.storage.introspect("bucketName")).toBe("local-bucket");
   });
 });